requirements.txt
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
deprecation==2.1.0
Django==5.1
frozenlist==1.4.1
gotrue==2.7.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
multidict==6.0.5
packaging==24.1
pillow==10.4.0
postgrest==0.16.10
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.0.1
requests==2.32.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.1
supafunc==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websockets==12.0
wheel==0.44.0
yarl==1.9.4
